// workspaces/server/src/types.ts
export interface RedisConfig {
  port?: number
  host?: string
  family?: number
  password?: string
  db?: number
}
// ...
export type WakaParams<T> = T
export type WakaBody<T> = T

export interface WakaRequest<WakaBody, WakaParams> extends Request {
  body: WakaBody
  params: WakaParams
}
// ...
export interface WakaVehiclePosition {
  v_id?: string
  latitude: number
  longitude: number
}